requirements.txt
asgiref==3.8.1
colorama==0.4.6
coverage==7.6.7
Django==5.1.2
django-cors-headers==4.5.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
iniconfig==2.0.0
packaging==24.2
pillow==11.0.0
pluggy==1.5.0
psycopg2==2.9.10
pygame==2.5.2
PyJWT==2.9.0
pytest==8.3.3
sqlparse==0.5.1
tzdata==2024.2